Separate src args when compiling asm with clang-cl
Currently, when compiling with clang-cl, we separate the source from
the arguments by inserting a `--`. This avoids compilation errors due
to conflicts between source paths and arguments (especially on macos,
where path will often begin with /Users, but /U is a clang-cl flag).

However, this separator is disabled when we are compiling assembly for
msvc targets, as those will usually be compiled through `ml.exe`, which
doesn't understand this separator. Problem is, this is not always the
case: Only `.asm` files are compiled using `ml.exe`. Any asm using a
different extension will still go through the normal C compiler
(clang-cl), and hence we need to provide the separator in that case.